third_party/blink/web_tests/external/wpt/css/css-anchor-position/position-try-rule-caching-002-expected.txt

This is a testharness.js-based test.
[FAIL] Modifying one sheet containing @position-try rule doesn't affect another sheet
  assert_equals: expected "right top" but got "right bottom"
Harness: the test ran to completion.
